Added documentation, cleaned up code, updated home page table
