+++ src/_scss/pages/homepage/visualizations/geo/mapList.scss
@@ -26,14 +26,18 @@
     .map-table {
         // implement the library style
         @import 'lib/ibTable/ibTable';
+
+        .ibt-table-container {
+            border: 1px solid $color-gray-light;
+            background-color: $color-white;
+        }
+
         max-width: rem(1025);
 
         margin-top: rem(15);
         margin-bottom: rem(15);
         display: block;
-        border: 1px solid $color-gray-light;
-        // placeholder background for when the user scrolls really fast
-        background: repeating-linear-gradient( $color-white, $color-white 40px, $color-gray-lightest 40px, $color-gray-lightest 80px);
+
         &.no-results {
             border-bottom: none;
         }
