+++ src/js/components/homepage/visualizations/geo/MapList.jsx
@@ -34,18 +34,14 @@ export default class MapList extends React.Component {
             tableWidth: 0,
             tableLeft: 0,
             tableRight: 0,
-            columns: [],
-            contentWidth: 0,
             sort: {
                 field: 'state',
                 direction: 'asc'
             },
-            sortedData: [],
-            renderHash: ''
+            sortedData: []
         };
 
         this.setTableWidth = this.setTableWidth.bind(this);
-        this.rowClassName = this.rowClassName.bind(this);
         this.changeSearchOrder = this.changeSearchOrder.bind(this);
     }
 
@@ -81,14 +77,6 @@ export default class MapList extends React.Component {
         });
     }
 
-    rowClassName(index) {
-        let evenOdd = 'odd';
-        if ((index + 1) % 2 === 0) {
-            evenOdd = 'even';
-        }
-        return `homepage-map-table-row ${evenOdd}`;
-    }
-
     sortData(data) {
         return orderBy(data, (item) => {
             const compareValue = item[this.state.sort.field];
@@ -103,63 +91,56 @@ export default class MapList extends React.Component {
     prepareTable(data) {
         // prepare the data
         const sortedData = this.sortData(data);
-        const renderHash = `map-list-sort-${uniqueId()}`;
 
         // prepare the columns
         let contentWidth = 0;
         const columns = [];
 
         tableColumns.order.forEach((column, index) => {
+            const columnX = contentWidth;
             contentWidth += tableColumns.widths[column];
 
-            let isLast = false;
-            if (index + 1 === tableColumns.order.length) {
-                isLast = true;
-            }
-
             const tableCol = {
+                x: columnX,
                 width: tableColumns.widths[column],
-                name: column,
-                columnId: column,
-                rowClassName: this.rowClassName,
-                header: (
-                    <MapListHeaderCell
-                        label={tableColumns.labels[column]}
-                        column={column}
-                        defaultDirection={tableColumns.defaultDirection[column]}
-                        isLastColumn={isLast}
-                        order={this.state.sort}
-                        changeSearchOrder={this.changeSearchOrder} />
-                ),
-                cell: (rowIndex) => {
-                    let formattedValue = `${sortedData[rowIndex][column]}`;
-                    if (column === 'amount' || column === 'capita') {
-                        formattedValue = MoneyFormatter.formatMoney(sortedData[rowIndex][column]);
-                        if (sortedData[rowIndex][column] === 'N/A') {
-                            // handle N/A values
-                            formattedValue = 'N/A';
+                header: () => {
+                    const isLast = index + 1 === tableColumns.order.length;
+                    return {
+                        headerClass: MapListHeaderCell,
+                        additionalProps: {
+                            isLast,
+                            field: column,
+                            title: tableColumns.labels[column],
+                            defaultDirection: tableColumns.defaultDirection[column],
+                            changeSearchOrder: this.changeSearchOrder,
+                            currentSort: this.state.sort
                         }
+                    };
+                },
+                cell: (rowIndex) => {
+                    const isLast = index + 1 === tableColumns.order.length;
+                    let cellData = sortedData[rowIndex].get(column);
+                    if (index === 1 || index === 3) {
+                        cellData = MoneyFormatter.formatMoney(cellData);
                     }
-
-                    return (<ResultsTableGenericCell
-                        key={`cell-${column}-${rowIndex}`}
-                        rowIndex={rowIndex}
-                        data={formattedValue}
-                        dataHash={renderHash}
-                        column={column.columnName}
-                        isLastColumn={isLast} />);
+                    return {
+                        cellClass: ResultsTableGenericCell,
+                        additionalProps: {
+                            isLast,
+                            rowIndex,
+                            data: `${cellData}`
+                        }
+                    };
                 }
             };
 
             columns.push(tableCol);
         });
 
-        this.setState({
+        return {
             columns,
-            contentWidth,
-            sortedData,
-            renderHash
-        });
+            contentWidth
+        };
     }
 
     changeSearchOrder(field, direction) {
@@ -171,14 +152,21 @@ export default class MapList extends React.Component {
             }
         }, () => {
             // update the table
-            this.prepareTable(this.props.data);
+            if (this.tableComponent) {
+                this.tableComponent.updateRows();
+            }
         });
     }
     render() {
+        const calculatedValues = this.prepareTable(this.props.data);
+
         let inFlightClass = 'loaded-table';
         if (this.props.loading) {
             inFlightClass = 'loading-table';
         }
+
+        const visibleWidth = Math.min(this.state.tableWidth, calculatedValues.contentWidth);
+        const visibleHeight = Math.min(tableHeight, this.props.data.length * rowHeight);
 
         return (
             <div className="homepage-map-list">
@@ -196,15 +184,16 @@ export default class MapList extends React.Component {
                         marginRight: this.state.tableRight
                     }}>
                     <IBTable
-                        dataHash={`${this.state.renderHash}-${this.state.tableWidth}`}
-                        resetHash={this.state.renderHash}
+                        contentWidth={calculatedValues.contentWidth}
+                        bodyWidth={visibleWidth}
+                        bodyHeight={visibleHeight}
                         rowHeight={rowHeight}
                         rowCount={this.props.data.length}
                         headerHeight={50}
-                        width={this.state.contentWidth}
-                        maxWidth={this.state.tableWidth}
-                        maxHeight={tableHeight}
-                        columns={this.state.columns} />
+                        columns={calculatedValues.columns}
+                        ref={(table) => {
+                            this.tableComponent = table;
+                        }} />
                 </div>
                 <div className="map-list-citation">
                     <p>Sources: US. Census Bureau, Population Division</p>
