+++ src/js/components/sharedComponents/IBTable/managers/RenderQueue.js
