--- src/js/components/sharedComponents/IBTable2/IBTable.jsx
@@ -1,8 +1,0 @@
-/**
-  * IBTable.jsx
-  * Created by Kevin Li 12/8/16
-  **/
-
-import Table from './components/Table';
-
-export default Table;
